+++ libs/react/fpkit/src/components/badge/index.stories.tsx
@@ -1,10 +1,11 @@
 import React from "react"
+import '@shawnsandy/first-paint/dist/css/components/badge.min.css'
+import { BADGE } from "@geometricpanda/storybook-addon-badges"
 
 import {
   ComponentStory,
   ComponentMeta
 } from "@storybook/react"
-import { BADGE } from "@geometricpanda/storybook-addon-badges"
 
 import {
   within,
@@ -13,7 +14,7 @@ import {
 } from "@storybook/testing-library"
 import { expect } from "@storybook/jest"
 
-import Badge from "./fp-badge"
+import {Badge} from "./fp-badge"
 export default {
   title: "FPKIT Components/Badge",
   component: Badge,