+++ libs/react/fpkit/src/components/cards/card.tsx
@@ -6,7 +6,7 @@ export interface CardProps extends ComponentProps {
 }
 
 export const defaultStyles = {
-  padding: "var(--card-p, 1.6rem)",
+  padding: "var(--card-p, 2rem)",
   backgroundColor: "var(--card-bg, white)",
   boxShadow:
     "var(--card-shadow, 0 0 0.5rem 0.1rem rgba(0, 0, 0, 0.1))",
@@ -15,24 +15,26 @@ export const defaultStyles = {
   color: "var(--card-cl, black)"
 }
 
-// const cardRef = React.useRef<HTMLDivElement | null>(null)
-
-const Card = ({
+export const Card = ({
   elm = "div",
   styles,
   children,
   renderStyles = true,
+  dataStyle,
+  id,
   ...props
 }: CardProps) => {
   const stylesObj = renderStyles ? defaultStyles : {}
   return (
     <FP
       as={elm}
+      id={id}
       styles={{
         ...stylesObj,
         ...styles
       }}
       renderStyles={renderStyles}
+      data-card={dataStyle}
       {...props}
     >
       {children }
@@ -40,4 +42,4 @@ const Card = ({
   )
 }
 
-export default Card
+Card.displayName = "Card"