+++ libs/react/fpkit/src/components/dropdowns/dropdown.tsx
@@ -1,6 +1,6 @@
-import Details from "./dropdown-details"
-import Summary from "./dropdown-summary"
-import { ComponentProps } from "../../types"
+import { Details } from './dropdown-details'
+import { Summary } from './dropdown-summary'
+import { ComponentProps } from '../../types'
 
 export interface DropdownProps extends ComponentProps {
   title: string
@@ -9,19 +9,19 @@ export interface DropdownProps extends ComponentProps {
 }
 
 const defaultStyles = {
-  display: "flex",
-  placeContent: "flex-start",
-  placeItems: "center",
-  gap: ".7rem",
-  border: "var(--summary-border, solid 1px #ccc)",
-  transition: "all 0.2s ease",
-  backgroundColor: "var(--summary-bg, whitesmoke)",
-  padding: "var(--details-pd, 1.2rem)",
-  minWidth: "var(--summary-min-w, 80vw)",
-  maxWidth: "var(--summary-min-w, 80vw)"
+  display: 'flex',
+  placeContent: 'flex-start',
+  placeItems: 'center',
+  gap: '.7rem',
+  border: 'var(--summary-border, solid 1px #ccc)',
+  transition: 'all 0.2s ease',
+  backgroundColor: 'var(--summary-bg, whitesmoke)',
+  padding: 'var(--details-pd, 1.2rem)',
+  minWidth: 'var(--summary-min-w, 80vw)',
+  maxWidth: 'var(--summary-min-w, 80vw)',
 }
 
-const Dropdown = ({
+export const Dropdown = ({
   styles,
   children,
   summary,
@@ -31,15 +31,12 @@ const Dropdown = ({
 }: DropdownProps) => {
   const stylesObj = renderStyles ? defaultStyles : {}
   return (
-    <Details
-      styles={{ ...stylesObj, ...styles }}
-      onToggle={toggle}
-      {...props}
-    >
+    <Details styles={{ ...stylesObj, ...styles }} onToggle={toggle} {...props}>
       <Summary>{summary}</Summary>
       {children}
     </Details>
   )
 }
 
-export default Dropdown
+// export default Dropdown
+Dropdown.displayName = 'Dropdown'