+++ libs/react/fpkit/src/types/shared.ts
@@ -1,6 +1,8 @@
 export interface ComponentProps {
   children?: React.ReactNode
   renderStyles?: boolean
+  id?: string
   styles?: {}
   classes?: String
+  dataStyle?: string
 }